requirements.txt
livekit
livekit-agents>=1.0.0
livekit-plugins-deepgram>=1.0.0
livekit-plugins-elevenlabs>=1.0.0
livekit-plugins-google>=1.0.0
livekit-plugins-openai[vertex]>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-rag>=1.0.0
python-dotenv~=1.0
aiofile~=3.8.8
twilio
websocket-client